Stroke on Remove brushes now spans outside the volume + Added check for Remove brush sides not spanning across both X and Y axis

+++ JERC/Models/OverviewPositionValues.cs
@@ -31,8 +31,8 @@ namespace JERC.Models
         public int outputResolution;
 
         //  used to align the image to center if the width != height
-        public float brushVerticesOffsetX;
-        public float brushVerticesOffsetY;
+        public int brushVerticesOffsetX;
+        public int brushVerticesOffsetY;
 
         //public float offsetPercentageX;
         //public float offsetPercentageY;
@@ -65,8 +65,8 @@ namespace JERC.Models
             paddingPercentageX = (float)paddingSizeX / (float)outputResolution;
             paddingPercentageY = (float)paddingSizeY / (float)outputResolution;
 
-            brushVerticesOffsetX = width < height ? ((height - width) / 2) + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountWidth / 2) : 0 + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountWidth / 2);
-            brushVerticesOffsetY = height < width ? ((width - height) / 2) + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountHeight / 2) : 0 + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountHeight / 2);
+            brushVerticesOffsetX = (int)(width < height ? ((height - width) / 2) + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountWidth / 2) : 0 + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountWidth / 2));
+            brushVerticesOffsetY = (int)(height < width ? ((width - height) / 2) + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountHeight / 2) : 0 + jercConfigValues.strokeWidth - (radarSizeMultiplierChangeAmountHeight / 2));
 
             //offsetPercentageX = (brushVerticesOffsetX / width);
             //offsetPercentageY = (brushVerticesOffsetY / height);
